# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "代理管理服务",
        "title": "代理管理服务",
        "contact": {},
        "version": "V0.1"
    },
    "paths": {
...
        "/api/v1/userGroup": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取用户组列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户组信息管理"
                ],
                "summary": "获取用户组列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "备注",
                        "name": "mark",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "组名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "总数  由服务器返回回去",
                        "name": "total",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/controller.Full"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加用户组信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户组信息管理"
                ],
                "summary": "添加用户组信息",
                "parameters": [
                    {
                        "description": "添加用户信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repository.CreateUserGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            }
        },
        "/api/v1/userGroup/state/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改用户组状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户组信息管理"
                ],
                "summary": "修改用户组状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            }
        },
        "/api/v1/userGroup/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "通过编码获取用户组信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户组信息管理"
                ],
                "summary": "通过编码获取用户组信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Full"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.UserGroupInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改用户组信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户组信息管理"
                ],
                "summary": "修改用户组信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改用户信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repository.ModifyUserGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除用户组",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户组信息管理"
                ],
                "summary": "删除用户组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.Base": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                },
                "requestDatetime": {
                    "description": "请求时间",
                    "type": "string"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string"
                },
                "responseDatetime": {
                    "description": "返回时间",
                    "type": "string"
                }
            }
        },
        "controller.Full": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据"
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                },
                "requestDatetime": {
                    "description": "请求时间",
                    "type": "string"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string"
                },
                "responseDatetime": {
                    "description": "返回时间",
                    "type": "string"
                }
            }
        },
...
        "model.UserGroupInfo": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "编码",
                    "type": "string"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "组名称",
                    "type": "string"
                },
                "state": {
                    "description": "状态 0 停用 1 启用",
                    "type": "integer"
                },
                "sys": {
                    "description": "系统生成 0 否 1 是",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "修改时间",
                    "type": "string"
                }
            }
        },
...
        "repository.CreateUserGroup": {
            "type": "object",
            "properties": {
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "组名称",
                    "type": "string"
                }
            }
        },
...
        "repository.ModifyUserGroup": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "编码",
                    "type": "string"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "组名称",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}